requirements.txt
pip
discord.py
moviepy.editor